portail_gestion/requirements.txt
certifi==2025.1.31
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-sslserver==0.22
djangorestframework==3.16.0
mysqlclient==2.2.7
pipenv==2024.4.1
PyJWT==2.10.1
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pypiwin32==223; sys_platform == 'win32' 
sqlparse==0.5.3
virtualenv==20.30.0
Werkzeug==3.1.3
python-dotenv
gunicorn==20.1.0
